Move Medium interfaces to own file
